# docs/openapi/pecr.v1.yaml
openapi: 3.1.0
info:
  title: PECR API
  version: 1.0.0
  description: |
    Versioned API contract for PECR v1.
servers:
  - url: http://127.0.0.1:8080
    description: Gateway
  - url: http://127.0.0.1:8081
    description: Controller
paths:
...
  /v1/policies/capabilities:
    get:
      summary: Describe safe and useful ask categories for the current principal
      operationId: getPolicyCapabilities
      security:
        - PecrLocalPrincipal: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XPecrRequestId"
      responses:
        "200":
          description: Safe ask catalog derived from current policy narrowing hints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafeAskCatalog"
        "401":
          description: Missing or invalid principal credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Source unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/policies/simulate:
    post:
      summary: Simulate a policy decision
      operationId: simulatePolicy
      security:
        - PecrLocalPrincipal: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XPecrRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicySimulationRequest"
      responses:
        "200":
          description: Policy simulation decision payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicySimulationResponse"
        "400":
          description: Invalid request payload or fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid principal credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Source unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "504":
          description: Policy engine timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    PecrLocalPrincipal:
      type: apiKey
      in: header
      name: x-pecr-principal-id
      description: Local mode principal identity header (`PECR_AUTH_MODE=local`).
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC bearer token (`PECR_AUTH_MODE=oidc`).
    PecrSessionToken:
      type: apiKey
      in: header
      name: x-pecr-session-token
      description: Opaque session capability token returned from `POST /v1/sessions`.
  parameters:
...
    XPecrRequestId:
      name: x-pecr-request-id
      in: header
      required: false
      schema:
        type: string
      description: Optional caller-provided id used for log correlation.
...
  schemas:
...
    TerminalMode:
      type: string
      enum:
        - SUPPORTED
        - INSUFFICIENT_EVIDENCE
        - INSUFFICIENT_PERMISSION
        - SOURCE_UNAVAILABLE
    ClientResponseKind:
      type: string
      description: |
        Optional client UX hint for `/v1/run` and error responses.
        Absent means the response is a normal supported answer without special handling.
      enum:
        - blocked
        - ambiguous
        - source_down
        - partial_answer
...
    PlannerIntent:
      type: string
      enum:
        - default
        - structured_lookup
        - structured_aggregation
        - evidence_lookup
        - version_review
        - structured_evidence_lookup
        - structured_aggregation_evidence
        - structured_version_review
...
    SafeAskCapability:
      type: object
      additionalProperties: false
      required:
        - capability_id
        - intent
        - title
        - description
      properties:
        capability_id:
          type: string
          minLength: 1
        intent:
          $ref: "#/components/schemas/PlannerIntent"
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        examples:
          type: array
          items:
            type: string
            minLength: 1
        scope_labels:
          type: array
          items:
            type: string
            minLength: 1
        view_ids:
          type: array
          items:
            type: string
            minLength: 1
        field_labels:
          type: array
          items:
            type: string
            minLength: 1
        dimension_labels:
          type: array
          items:
            type: string
            minLength: 1
        source_scopes:
          type: array
          items:
            type: string
            minLength: 1
        document_hints:
          type: array
          items:
            type: string
            minLength: 1
    SafeAskCatalog:
      type: object
      additionalProperties: false
      properties:
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/SafeAskCapability"
        suggested_queries:
          type: array
          items:
            type: string
            minLength: 1
...
    PolicySimulationRequest:
      type: object
      additionalProperties: false
      required:
        - action
        - params
      properties:
        action:
          type: string
          minLength: 1
        params:
          type: object
        policy_snapshot_hash:
          type: string
          pattern: "^[a-f0-9]{64}$"
          description: Optional override for policy snapshot hash; when omitted the gateway forwards no snapshot hash.
        policy_bundle_hash:
          type: string
          pattern: "^[a-f0-9]{64}$"
          description: Optional override for policy bundle hash; when omitted the gateway default bundle hash is used.
        as_of_time:
          type: string
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
          description: Optional RFC3339 UTC timestamp; when omitted the gateway default as_of_time is used.
    PolicyDecision:
      type: object
      additionalProperties: false
      required:
        - allow
        - cacheable
      properties:
        allow:
          type: boolean
        cacheable:
          type: boolean
        reason:
          type: string
        redaction:
          type: object
        narrowing:
          type: object
    PolicySimulationResponse:
      $ref: "#/components/schemas/PolicyDecision"
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - terminal_mode_hint
        - retryable
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        terminal_mode_hint:
          $ref: "#/components/schemas/TerminalMode"
        retryable:
          type: boolean
        response_kind:
          $ref: "#/components/schemas/ClientResponseKind"
        what_failed:
          type: string
          minLength: 1
          description: Concrete explanation of the blocked or unavailable step.
        safe_alternative:
          type: string
          minLength: 1
          description: Suggested safe next step for the caller or user.
        detail:
          oneOf:
            - type: object
            - type: array
            - type: string
            - type: number
            - type: integer
            - type: boolean
            - type: null